+++ packages/in_app_purchase/in_app_purchase_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.3+3
+
+* Add `implements` to pubspec.
+
 # 0.1.3+2
 
 * Removed dependency on the `test` package.