+++ script/tool/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added Android native integration test support to `native-test`.
 - Added a new `android-lint` command to lint Android plugin native code.
+- Pubspec validation now checks for `implements` in implementation packages.
 
 ## 0.5.0